driver semcomedi: fix acq thread sometimes not starting
If acquisition is requested just after it was stopped, there was a little
chance that it would not be restarted again.
=> increase the checks.
In the future, we'd be better off with an acquisition thread constantly running,
and listening to the DataFlow requests via a queue of commands.

This fixes timeouts during acquisition of the SPARC with small exposure time.