acq stream MD: fix MD_POS of assemble data
For the SEMMD stream, the translation was set before the resolution, which prevented
it from picking up the right value (because the e-beam driver would force the translation
to stay at the center).

For the SEMCCD stream with fuzzing, the metadata computation was wrong due to giving the
wrong image resolution.