gui model: ensure that the first focus move is applied at least 50ms after beginning
Initialy the idea was to immediately apply the first move.
But when doing a small quick focus move, there are still at least two requests,
when moving and when releasing the mouse, so it causes two small moves.

To avoid lots of small moves, which can be slow on closed loop devices, we always
wait a bit if there is move requests coming.