model: fix .parent not being set on the ComponentProxy
It was trying to be clever by using a weakref, but as the proxy has no ref to the parent,
it was immediately garbage collected.

So just keep a ref (to the parent proxy).

Also remove the ability to change the parent remotely. Now the parent really just means the "creator", so
there is no need to change it. And remove the instances that were doing just that.