util img: handle better data with very large data range
Data type like uint32 would get histogram so compressed that it doesn't keep the information on the min/max.

So clamp the data range in such case.

Also don't use the shortcut of bincount if drange != 0, as bincount starts always at 0.

Also fix problem of compactHistogram trying to extend the array.