gui cont streams: automatically update the resolution VA based on scale/binning
When the scale or binning are modified, the resolution changes too... at least if the
hardware hears about it. If it's a local VA and the stream is not playing, nothing will happen.
In some cases, that caused the image to be cropped when not requested.

So we now update the resolution based on the scale or binning.
For now we always assume the user wants a full FoV.
That might sometimes be not true on the SECOM, where we allow the user to change the resolution
manually, but the best way to solve this is to add a "crop ratio" selection.

This might already allow to remove some hack in the stream. And maybe local VA resolution should
never been sent to the hardware as-is (only when computed from the ROI).