driver pmtctrl: take lock on the tty
To be sure not to collide with odemis-relay when a unit is just connected,
we need to take a lock. This can typically happen if (re)plugging in the device when
odemis starts.

Also clean up a bit the left over of serial number.
And let the test case be TEST_NOHW aware.