gui comp legend: Ensure we really cannot call wx from outside the main thread
changing the tooltip could use wx, so make sure it is called from the main thread.
Also simplify the way the tooltips are set, by using the normal wxPython way.

Also some minor optimisations.

With this, the GUI has not crashed yet... but that might be just a matter of time.