Add .focuser info to the streams, and use that instead of per view in the GUI
The problem of the focuser is that it's a matter of a stream: it affects a specific path from an emitter to a detector.
So it didn't make so much sense to have it per view, as a view can have streams with various emitter/detector streams.
In particular, that showed on the DELPHI combined views which always changed the optical focus, even when the SEM stream
was playing.
(Note that technically, there could be _several_ focusers on a path emitter->detector. For instance, on CL acquisition.
But for now we consider the user will always expect to modify only the "major" one.)

So now allow to store the focuser info on the stream, and use it in the GUI.
That means that the viewport and canvas must be a bit more dynamic and cannot rely just on the fact the view has a focuser or not.
Every time the stream changes, the focuser can have changed.

Also slightly modify the way the focus move is requested from the canvas, in order to be able to support (later)
a different pixel->meter conversion based on the stream/focuser type.