Add support for depthOfField estimation on optical cameras
The goal is to provide enough information to know what is a meaningful focus step.

Support adding numerical aperture and refractive index information to a lens.
Add depthOfField VA to all the cameras, and automatically update the info based on the metadata received from the lens.
Connect the new lens VAs to the new camera metadata in the mdupdater (and also fix the updater)