driver blinkstick: rename whiteled to blinkstick and some improvements
The hardware is called blinkstick, so call the driver blinkstick. The class is still WhiteLed.

Also add info on hwVersion, and give the potential to the user to pick a device by serial
(although the current version of the low-level driver has a bug for now).

Add .spectra VA for compatibility with the rest of odemis drivers. Probably will need to
allow the user define the spectra of the led on the hardware.
Add a .scan() method.

Remove the simulator as it's the same as the simulated.Ligth component.