gui cont microscope: on Delphi loading, explicitly set the focus of the navcam
Now that it was removed from the phenom driver, do it from the GUI.
Also slightly change the value to get closer from the current value.
The main goal is that it can later be more flexible: either do an autofocus, or rely on some predefined value per phenom and sample holder.