acq stream: improve rounding of the drange based on data value
When drange is based on data value, instead of always rounding the width of the drange to a power of 2,
also check if a drange between 0 and max rounded would also work.
In practice, that means we more often end up with a range 0->power of 2, which is slightly more
user friendly (at least more nerd-friendly).
Especially that avoids ending up with a drange that has a bigger max than the original data.