rearange icon files for the GUI
Update the standard icon and the viewer icon to match together.
Now they are stored as png in the image directory, and all the icons
derived are automatically generated by a script (util/generate_icons.sh)

Also update the GUI to use the right icon based on whether is in standalone
(viewer) mode or full feature mode.