driver tmcm: Begin of support for TMCM 6110
Accept less or more than 3 axes. Not only allow 6 axes on the 6110, but also allow less axes
if less axes are connected.

Allow to select a controller by its "address" (which is typically set via the DIP on the board).
That permits to select the right controller if several controllers are connected. That will
happen on the SPARCv2.