driver tmcm: Add referencing support to the simulator
So now we can drop the fake referencing.
Also rename the "LS" reference procedure to standard, and make it use the configuration info.
Also reorganise a bit the loading of the configuration.
Also start support for turning on/off the reference switch before/after the referencing (will be need on SPARC2)