driver phenom: read actual spot size value at init
The phenom used to not support reading the spot size when unloaded,
so we were setting some (almost) random value.

Now that the Phenom supports reading the spot size at any time,
read the real value/range at init, so that we immediately show something
meaningful. That especially avoids having a strange 2.8xxx value in the GUI.