gui cont tabs: Connect ChamberTab
All the basics of the chamber tab should be there.

The main thing missing is the progress bar. => Still need to create a wrapper for the
standard Future to be seen as a ProgressiveFuture.

Also rename .notify()/clear_notificition() to highlight(), so that it's
clearer what it does.

Also change the step size of the TMCM in the simulator to make the mirror transition faster.