dataio: rename find_fittest_exporter() to find_fittest_converter() and allow to specify read or write
The function used to only work for writting files.
It is now also used for finding the best converter for reading.
However, the user need to be able to explicitly indicate the purpose in order to work
with converters that support just one direction.

Simplify the function and support an explicit purpose.

Also rename the function to make the purpose clearer.