gui: don't completely fail if a spectrum is flat
If the range of the data in the spectrum was empty (eg: all the data is clipped to the maximum value),
bad things would happen... including a freeze of the GUI.

So now we handle this better. At least, we don't try to fit an infinite amount of ticks in the legend.
Also avoid the division by 0 in the range computation.