driver semcomedi: workaround writter thread not stopping
Since commit 2de30d4ebf "change the way the acquisition thread is managed",
cancelling a big acquisition and immediately starting a new one was blocking.

That's because it avoided to move the ebeam to rest position if a new acquisition was due
(as an optimisation). However, setting to rest position was (unknowingly) helping to
actually unblock the write() of the writer thread.

So now the writer thread would never end, and block every new acquisition.

To workaround that, we write to rest position after every cancellation.
Note that it seems it can still happen that the write doesn't unblock.
We try to detect that rare case, and raise an exception so that the whole
device is reset and acquisition can restart anyway.