gui tabs: don't start playing align streams at init of SPARCv2
Due to the CallAfter, the VA callback on align_mode would be called after
show(False), and actually start the streams.

Just don't call it at init.
Do the same for the SPARC, though it was not a problem as the Alignment tab is
the default tab.

Also small clean ups