acq stream: Fix sync acquisition failure to assemble data if timeout occured
We now try harder if there was a timeout from the CCD.
However, the SEM data (likely) arrived. So after a re-trial, we end up with too many SEM data.
That could have happened too if the data had been received too fast, but much less likely.

So now, we make sure to trim the last SEM data if it's already received.