acq stream: allow to get the min/max of the raw values
Pretty kludgy, but for now it seems it could help the user if we displayed these values.
So display them, and see if we keep that or not. Maybe we should just display the histogram?

Also allow to retrieve the raw CCD image. Also for potential use in the GUI.

Also fix the histogram trying to be computed for the live spectrum.