tmcm: Workaround bug in TMCM-6110 in a different way and allow to reset the memory
Instead of explicitly reseting the parameters at driver init, put the commands in
a routine which is called at start. The advantage is that the routine is called just
at the right moment, after turning on the power supply. This is the recommended way
from Trinamic.

Add a --reset option to reset the memory of the board to the default.