gui: connect mirror overlay to polePosition VA
Add methods to the mirror overlay to change the size, and to use a VA.

Make sure it uses the Phyisical coordinates (not the World coordinates).

Introduce new polePositionPhysical VA to store the pole position in physical coordinates.

Link/convert back and forth the polePositionPhysical VA of the tab and polePosition VA of the lens.