acq stream: make the .tint VA generic to all VAs
Instead of having tint only part of the FluoStream, allow to tint any stream.

Handy for tinting the SPARCv2 line alignment stream.

Also simplifies the code, as we don't need to pass the tint as argument of ._updateImage() anymore

Also need to update the GUI to only display the tint selection button on FluoStreams.