acq stream MoI: handle when the MoI is NaN
the min() and max() of a an array containing a cell with NaN is NaN.
MoI is NaN if the image is black.
We care more about the real values, so use the special numpy functions that handle this
correctly.

Also turn the pixels with NaN red too, to indicate it's probably not interesting.