gui overlay line selection: force a pixel selection whenever the line is selected
Right now, if the pixel is already picked, and somewhere else, it's confusing
for the user because it's outside of the line, but the spectrum is displayed.
So default to the starting pixel. Could be slightly more clever when slightly moving the line,
but hopefully it's not a big deal.