gui: use the Parabolic Mirror Overlay aslo for the SPARCv1
The goal is to simplify the canvas code.

The drawback is that the horizontal line might be located at a bit different place
than previously (depending on the CCD size). So systems currently with a good pole position
will need to be recalibrated.
=> Acquire an AR image in full CCD size with good signal, open the file in
hdfview to see the raw data, and locate the hole in the mirror. Report that position
in the microscope file.