cont tabs sparc2 align: only save the mirror and lens position when tab is shown
Sometimes the Show(False) function is called even if the tab is already not shown.
(eg, when closing the GUI)
That could leave the favourite positions as incorrect.
We could check explicitely for such a case, but it's probably better to just change the
behaviour to be closer to what we expect.