gui powermate: Don't fail if python-evdev is not present
Python-evdev only makes sense on Linux. We cannot depend on this being always present.

So move the imports inside the code, and just raise LookupError or NotImplementedError
if it's not available.
Make the GUI main also not afraid if Powermate fails.

Also some various small clean ups to the powermate code.