driver andorcam3: don't send data if no one to receive it
It could happen that a subcription is stopped, which stops the acquisition.
But during the 0.1 s of time the acquisition thread is waiting for the hardware,
the settings are changed and a new subscription arrives, and the hardware provides the data.

In that case, the new subscriber would receive the data while the settings hasn't
been updated. It's not purely an error according to the spec, but it's quite a
strange behaviour and better to avoid.