driver andorshrk: Fail horibly instead of blocking forever if SDK is confused at init
The SDK 2.100 can block if it's initialising while the Shamrock is initialising, or just because.
In that case, for now we get stuck forever (not giving any user feedback).

Unfortunately, because we are in C code, it's pretty hard to come back to Python and nicely raise an Exception.
So instead, just die misarably, it's still a bit better than being a zombie for the rest of the time, right?