driver spectrometer: only update the binning/resolution setting when acquiring
The goal is to allow to access the CCD used by the spectrometer also as a standard CCD.

Also modify the getPixelToWavelength() API to get pixel size and resolution in its arguments.
That avoids the spectrograph to need a CCD, and also allows to get that info even if the CCD
actually hasn't been updated (yet).

It now important to acquire spectrometer data via the "spectrometer", not the "sp-ccd".