andorshrk: try harder with hw access lock
Still locking when calling GetCalibration().
Some calls were not protected, so protect them to be really sure.

Also change the lock to be reentrant, because that matches more the meaning of the lock.
Currently, in Python 2.7, a RLock is about 10x more costly than a Lock
(ie, about 1 µs instead of 100ns to acquire/release). In Python 3, they are almost the
same cost, and anyway, it's already worthy to make sure we avoid deadlocking...