acq stream sync: revert the way to synchronise the ebeam
Turns out that I had overlooked an important behaviour when using the event synchronization
for the ebeam: it'll not stop the current acquisition.

That causes the settings (aka translation) to only be updated on the end of the current acquisition.
This can be after the next CCD acquisition started, which is bad.

So revert to the old good time.