gui tabs sparc2 align: only enable the fiber align buttons when ready
As long as the spec-selector/fiber-aligner X axis is moving, the user shouldn't
try to move it too. So disable the buttons, until the optical path manager is done.

As setPath() doesn't return a future, we implement our own thread runner.
Hopefully, can be simplified once it returns a future.