gui streams: ensure the resolution of the CCD is full when forcing the binning to 1x1
At the moment the AR stream starts playing after creation, the code linking binning to
resolution is not yet connected. So it could end up with a good binning of 1x1, but
still have a spectrum-like resolution.

=> force resolution in such cases

Also make sure that no rounding happens when displaying the resolution of the spectrometer.