gui cont tabs sparc2 align: Force the AR image to be at 0,0
It's typically at 0,0 because there is no stage.
However, if there is a stage, the CCD data gets the current position.
For alignment, that doesn't make sense.
ContentView should work around that, but on very fast refresh rates, it doesn't work so well.
It's simpler to just force the position to 0,0.