driver andorcam2: allow to select a device by its serial number
When multiple devices are connected, we need to rely on something stable to
identify them (as device number can change after each run). Use the serial number.
The little drawback is that to read the serial number, the device must be initialise.
This means that we can end up initialising _all_ the devices to just use one.

Also add support for iDus InGaAs and iKon.