gui cont tabs: if the view HFW is linked to the Hw HFW, don't fit on first image
The purpose of 'fit on first image' is to adapt the view HFW to the hardware,
even if we don't know the hardware HFW beforehand.

That trick is not needed when we actually get explicitly that info from the hardware.
It makes things even more weird if the Hw HFW has just been changed (after starting
the image acquisition). And it slows down a bit the GUI at init.