driver pigcs: Handle E-861 which answers HLP? with a line that looks like an empty answer
Nicely, that controller doesn't follow the protocol, as it answers HLP? like this:
'4 HLP?\n'
'0 4 \nThe following commands are valid: \n....\n'

Notice how the first line is entirely empty, which caused the new code to think it was the
end of the answer. We would expect two spaces, or no \n.

=> If there is still stuff after, concider it as multi-line anyway.