gui comp: optimise MarkingLineOverlay when moving the mouse
With the BarPlotCanvas, whenever the v_pos VA changed, it was snapped and changed back.
Not only was snapping slow, but also it caused another computation, because the VA was changed again.

=> Inverse the logic, and let the overlay remember the exact value selected,
and it will snap to the real data only when actually displaying (which happens less often
than the mouse moves).