acq stream: more tweak with _updateDRange
From testing, spectra with 32 bits were not correctly handled.

Make sure to limit the drange using the data all the time it could be worthy.
Also factorize the code with the new method _guessDRangeFromDetector()

Also call _updateHistogram() without raw[0] if raw is defined.
That allows the spectrumStream to compute correctly the histogram, limited
by the bandwidth even at init.