gui: small improvements on Delphi manual calib
Make sure the manual calibration is run without block the GUI.
It used to work only if gnome-terminal was already running.

Share the code to run the manual calibration.

Also simplify a bit the recalibration menu handling (don't involve the tab controller).

Update the user text for the recalibration dialog.