gui delphi: cancel faster the delphi automatic calibration
The cancel() was return true, but actually the future just kept going on.

That was preventing Odemis from successfully closing, and also would make lots of weird
behaviour on the GUI if still used.

To be even better, we should also cancel the sub-future, so that it stop really just
when requesting it.