acq path: don't rely on special .positions roattribute and reduce referencing
.axes.choices contains the same as .positions, but it's generic => use this.

Also try to reduce the references kept after a move, hoping that more futures will be deref'd.

Eventually, the main issue is that optical path manager doesn't get deref'd because
the executor keeps a ref to it. It's solved in the new version of futures, so
we'll just wait until we get that new version.