driver pigcs: add ability to do target position move updates
Only supported for the RelCL (iow, C867 and E861 in closed loop).
That's because in open loop the controller is not able to tell us how many steps it has moved,
nor is it able to add up steps. So it'd be quite difficult to well add up (relative) moves.
We don't need (yet) updates on the AbsCL, so it's not implemented there, altough it should work as well.

Every move always starts the same, but while waiting for the move to be over, we also check
if another move is queued and is an update of the current move.

Also optimise the start of the moves to be as short as possible, to reduce latency.
On tests, we can do a new relative move (on one axis) every 30 ms.