driver actuator MultiplexActuator: support (a bit) update moves
Update moves are tricky as they need to break the current move.
The multiplex actuator doesn't even support cancelling the moves.
In addition, in case of several children we'd need to somehow detect if the
update move would really update all the children, to keep the move synchronised.

For now, only support such a move when it's obvious: there is only one child.
=> if only one child we now directly pass the move to the child, without using our own queue.
That has also the advantage to slightly improve the speed of such relatively typical use-case.