gui cont tabs: add comment about mirror image looking a bit bigger in SPARCv1
After checking the discrpancy between the new display and the old one, it seems
the main difference is that the main circle is 9% bigger. That's because
in the old version, the circle was concidered to have a radius of 10.5mm instead
of 11.5mm. It shouldn't matter too much as that border cannot be used for
alignment and the center of the circle was at the correct position.