driver andorcam3: discard old images which are on the hardware queue
If the hardware generates images faster than the driver can process,
the images will accumulate on the hardware queue. First, it will cause
latency in the display (if the user is looking at the images live).
Then, if the accumulation continues the hardware queue will get full and
the camera stop the acquisition causing a TIMEDOUT(?!) error.

So we check if there are any newer image already available. Note that in some
cases that doesn't work, apparently due to a bug in Andor SDK3' WaitBuffer(...,0).
Still, it helps in some cases.

Note that this whole situation happens very rarely as any relatively beffy computer
(as we have) can handle the maximum data rate that the Zylas can transmit over USB3.

Also, only do so if _not_ synchronised. When synchronised, there is no reason the hardware queue builds up anyway,
and the user likely want to get all the images.