gui xrc: use a bit less space for secom alignment quick tips
The log button doesn't show up if the window is not heigh enough, which
would happen on just a full HD screen.

The best would be to force that button to always be aligned at the bottom, but I don't know how to convince the sizer to do so.
At least, use less space for the doc display, so the log button appears more often.