driver pigcs: don't ask controller whether it's moving if we know it's not
When updating the position, the open loop controllers would be asked whether they are moving.
If no move is supposed to be going on, for sure it's not.
That slows done position update, and in turn that can cause garbage data
when the position update happen during a move.

=> Don't ask.

We still need a better way to make sure that position update doesn't interfer with the current move.