driver omicronxx: several small improvements
Ensure the device will start with light off when turned on manually.
As we only use the temporary power, if another value was set in memory, it could
automatically start with light when turned on. Not so much an issue if the device
is turned on by odemis, as it's immediately turned off, but a little dangerous
if the user turns on the device manually.

Send a HwError if some of the interlock is active. Some interlock can actually be
changed without the device being power cycled, but that forces a message at boot,
to make sure the user _sees_ the reason why there is no light.

Force the external modulation off. If the modulation was set on for some reason,
but no power controls the device, the laser would stay off without much explanation
for the user. If actually some user need that modulation, we can still update the driver
to support some explicit init arguments.