script monochromator scan: refactor and basic plugin
Use a stream to do the acquisition, so that it's possible to use the
acquisition manager, so that it's possible to easily share the code
between the plugin and the console version. Only drawback is that
cancelling halfway in the acquisition doesn't allow to save the
data acquired thus far anymore.

Plugin doesn't work yet, as the Plugin framework doesn't exist yet... but once it's here, it should work.