config: nicer password GUI when editing config or microscope file
The main problem of using sudo in a terminal, is that afterwards the terminal will stay around
and unity will think this is the main icon for the terminals.

The easiest way to use a GUI sudo is gksudo, but that doesn't work because for some
strange reason it passes a final empty argument to gedit that convinces gedit to
open a new empty tab.

=> use pkexec with special trick for being allowed to display stuff on the X11 display.