gui cont microscope: force external signal if SECOM SEM has no blanker
Now that the semcomedi driver supports TTL signal to indicate when it's scanning,
it could cause external signal to stop as soon as SEM acquisition stops.
For the SPARC, that's fine, but for the SECOM, it's troublesome as the SEM GUI
will likely cause scanning again, which is not good if an optical stream is playing.

=> Force the external signal to True when any stream is playing, to ensure the
beam gets at least parked.

Also now modify is_acquiring during the SECOM acquisition to ensure it's picked up
by the microscope controller.