gui model: slow down a bit more if > 2 futures are queued when moving focus
It can be that the hardware doesn't pick up the changes as fast as the GUI sends them.
In such case, the pending futures reach 3 or more... and the next focus move will be
queued again. In such a case, it's bettr to wait until all but the last moves have
been processed, so that we can send a bigger move next time.