gui plugin: reset logging level if it was changed
If a script (or plugin file) sets the logging level at the top of the file (as I like to do),
it will modify the logging level of _all_ the python program, including the GUI.
So check for this and reset the level after import.

It's still possible to change the logging level via a plugin, but in the plugin class itself.

Also report if a script has no plugin found.