gui acqmng: fix SECOM exposureTime shared for all optical stream in presets
A copy-paste caused the same value to be used for all the exposure times.

Also avoid keeping in presets the VAs which are not shown to the user.
There shouldn't be any reason to modify them, as the user wouldn't be aware.
Can even cause some behind-the-scene changes to not work.