driver pigcs: fix the timeout recovery procedure
Make sure that recovering on the E-861 (CL) when reading the position doesn't dead-lock.

Make sure that the whole recovery sequence doesn't get interrupted by other messages.
Often it works anyway, but sometimes recovery is needed because garbarge was received
on the line, and it could lead to several controllers trying to recover from that garbage.

Make sure that stop/halt commands are immediately reading the error. That slows down a bit
the stop procedure, but ensures that the error is not picked up when reading position.

Fix python error when doing a full recovery (ie, with a reboot).

Don't read position while axes are moving, to avoid the main cause of garbage on the line.

Increase message verbosity when a recovery is taking place.