driver pigcs: change auto_suspend PID=1,0,0 to 0,0,0
In some cases, even P=1 is too big and causes vibrations.
Maybe P=1 _could_ help in some cases to avoid drift, but that's too unsure.
The main goal is to avoid vribrations, so we need to be 100% they never happen.

Also, the encoder reading will still report the drift, so it's possible to compensate
the drift at higher level (by explicitely requesting a move)