gui plugin: run button callback in a separate thread
If the button callback is not run in a separate thread, it will hold the whole GUI from
updating until it's done. That's not a good behaviour for many functions that a user
could want to have when pressing a button. One example is to run an acquisition
and have the progress bar updated.

In order to allow that callback to modify the rest of the dialog, we need to
protect most of the methods to be running in the main thread.