dataio tiff: fix spectrum wavelength not saved if filter present
Old code thought that MD_OUT_WL is incompatible with MD_WL_LIST.
However, now that MD_OUT_WL can also be just a filter name, that's not
necessarily the case. For instance, when there is a filter, it's still
possible to save both the filter name and the wavelengths.