gui cont tabs: Fix cancelling fiber alignment mode causing the selector not to work
In case a new alignment mode was selected before the OPM move was ended, the subscribtion
to the spec-selector would be done when already in another alignment mode.
That would cause the active position of the spec-selector to be updated to the deactive position.
From then on, it'd be impossible to activate the spec-selector, as both position are identical.

Fix it by detecting at the end of the move that alignment mode has already been changed.