gui: handle more widgets destruction
Make sure that the acquistion window (SECOM/DELPHI) is fully destroyed when not in used.
That was causing quite some memory leaks.

Make sure that the VAConnector doesn't keep updating widgets if the widgets are gone.
That avoids in particular lots of useless calls whenever a hardware VA was changed when
many acquisition windows have been closed.

Simplify the stream panels refitting when they are removed.
In particular, when clearing the whole stream bar, it's not needed to refit the
streams for every stream, but only at the end.
Might speed up on Windows, when opening a new acquisition file in analysis tab.