acq stream sync: don't concider the CCD sent an image too fast is it was just 5% too fast
On some long acquisitions, users have seen failure to acquire because the CCD would take
419.99s to send the image instead of 420s.

The CCD clock can be going a little faster, and the whole expected acqusition time might
be a bit pessimistic. For those reasons it should be reasonable to have CCD image coming
up to 5% too early.