gui cont streams: Fix modifying spectrum bandwidth forcing it to 0 nm
In the analysis tab, on spectrum data, changing the bandwidth of the spectrum by
typing a number would always cause it to become 0.
Changing the center, would also cause the bandwidth to become 0.

That was due to a // instead of /, causing every computations to round down to 0.

Also allow explicitly bandwidth down to 0, as set in the range.