acq drift: clean up spaces and numpy warnings
No behaviour change.

The new numpy 1.11 complains about indexes being floats.
The floats were actually always positive integers (obtained with numpy.fix),
so it was not a big issue. Still, it's easier for the eyes and numpy,
to use an explicit division and get the same result.

Also remove the trailing spaces.