util img: mergeMetadata() shouldn't show warning about failure to correct pixel size if no pixel size correction
If trying to merge metadata which has no PIXEL_SIZE, mergeMetadata() would show a
warning that it's not able to correct it. That makes no sense (and clutter the log file)
if there is no PIXEL_SIZE_COR metadata either.

On the other hand, if there is such correction metadata, it should warn a bit more,
because it's quite a sign something went wrong.