install odemis-start: In case of error, just show the error message
Instead of showing the whole log, first only show the error message.
Still allow the user to see the full log if he wants. Seeing the full
log message is the default button, so that the user has to be at least
a little concious to discard the error message.