driver andorshrk: allow to return negative wavelengthes
For center wavelengthes near 0, there is no really good way to map all
the pixels to a wavelength.
We used to clamp the negative wavelengthes to 0. But even near 0, the values
are wrong.

Having lots of values for the same wavelength 0, was confusing other parts of Odemis.

As it should be obvious to any user that negative wavelengthes are wrong, we just
pass them as-is.