gui: attempt to reduce memory usage by using less pyro threads
It's quite ugly, as we rely on the fact that odemis.model is already loaded, and nothing else
will modify the Pyro config again. It'd be much better to have the Pyro servers correctly grow the number of threads,
or at least allow to select the number of threads at the server creation.

For now, that seems to help reducing the number of threads. For instance, running a sparc2 simulator,
we go down from ~200 threads to ~110 threads. Command to count threads:
ps -LlyF -u $USER | grep odemis.main.gui | wc -l

On 32 bits installations, that could be quite helpful.