driver andorshrk: be more careful when trying to force detector offset
Reading/writting the wavelength worked most of the time, but some times, at initialisation, we would get such error:
ShamrockError: Call to ShamrockSetWavelength failed with unknown error 20267: PARAMETER 2 INVALID
That's because the mirror grating doesn't support any wavelength but 0... however, it's possible
to read a different wavelength if the wavelength was set using a different grating.

So try to change the grating instead of the wavelength, as it also update the detector offset,
and it is hopefully less likely to have wrong parameter.
Also catch error in such case, and just show a warning instead of preventing Odemis from starting.