odemisd: add checks on YAML file for detecting lonely component and incorrect affects
Now can detect affects attributes which contain names of non-existing components.
For now, it only shows a warning, because there are quite a few (working quite well) microscope
files in the wild with such error.

Also detect components which are created, but absolutely not used. It's typically a
sign that this component is missing from the microscope children, or there was copy-paste error.

Also fix a couple of YAML files, based on what was detected.