util dataio: handle time data, in a very basic way
This is a quick hack to support showing time-correlated data (which we don't officially support yet).

The data is essencialy identical to a spectrum, with the time instead of wavelength, so convert the
time data into a spectrum data and pass it to the Spectrum Stream. This allows to visualise such data
in the GUI, with the only drawback that the plot legend indicates values in meters instead of seconds.

Most probably, the right way to do it on the long term is to generalise the StaticSpectrumStream to
3D stream (aka 'hyperstack'). The main change being the need to indicate the unit of the 3rd dimension.