driver andorcam3: attempt to reconnect to camera if it got disconnected
It can happen that the camera gets either disconnected (both physically, and from a kernel point of view),
or power-cycled.
Until now, as soon as it would happen, the driver would never be able to talk to the device anymore,
even if it comes back.
As the back-end is supposed to run long, it's good to handle such cases.

For now we handle only the 90% cases.
We try to reconnect to the camera only when an acquisition is requested.
In particular, we don't support:
 * reconnecting ahead of time (ie, as soon as it's connected again)
 * reconnection to read the temperature (temp just stops updating until next acquisition)
 * reconnection when changing a VA that directly talks to the hardware (=> they fail)

In practice, it seems to be relatively stable with a kernel 4.4 and SDK 3.11.30050