acq align autofocus: don't fail if time estimation cannot be properly done
If a detector doesn't have an exposureTime (eg, ETD) and not given any emitter with
a dwellTime, the autofocus would fail, just because time estimation cannot be done.

=> still allow autofocus, without a good time estimation.
Note that the code could be improved by using the time used for the first image
acquired to update the time estimation... but for now we don't care.