odemisd mdupdater: also push the lens metadata to sp-ccd
The sp-ccd get the raw 2D data of the spectrometer, and so it should be entitled
to exactly the same info as the ccd.

In practice, this allows the depthField VA to be more correct, and in turn this
means that autofocus has a lot more chance to work better/faster.