acq align autofocus: improve handling of focus reaching range limit
The autofocus code did check that the focus stayed in range, but didn't handle it well:
 * It would not use the cached focus value in such case
 * It could just keep 'boucing' between each range limit if the step was very big

This is the case of the focus of the SR193i, which has a range of less than 5 mm.

=> detect that the step is too big for the range, and reduce it.
IOW, this means it will get hint from the actuator range about the good step size.
=> if near a range limit, reduce the step size to avoid endless bouncing.

Also remove the 'et' argument which was unused.