gui sparcv2 alignment: use new autofocus spectrometer function and support external spectrometers with focus
Now that the autofocus spectrometer function is there, it's time to use it.
The main difficulty is to find the right components to pass to the function.

Also support spectrometers with a focus which are after the optical fiber.
In such case, the autofocus button is on the Fiber alignment, and there is
no slider to change the focus nor a stream to control.

Also automatically cancel the autofocus if the alignment mode is changed,
the same way as changing the tab would do it.