gui widgets: ProgressiveFutureConnector updates the tooltip text at the end of the future
If using a label, the end of the future shouldn't affect it, as in some case there is other
code that immediately uses that label for more information.
However, in the case of a tooltip, it wouldn't make much sense to use the tooltip to display
any other info than the progress. So indicate whether the future is done or cancelled, instead
of leaving "a few seconds left" forever.