model vattributes: avoid using != on numpy array
Numpy arrays do a element-wise comparison on !=, which is very costly, and especially
not needed if the previous value was not an array.

We already detected such case when the new value was a numpy array, but we also need
to check if the new value is not an array (eg, None), while the previous one was.

With new numpy version, that would cause warnings like this:
FutureWarning: comparison to `None` will result in an elementwise object comparison in the future.